- Moving ISettingsService to the Logic project. - Adding ISettingsService to the ShortcutProvider implementation.  - Checking if shortcuts are enabled before returning results.  - Adding DetectShortcuts to ShellViewModel.  - Adding DetectShortcuts to ShellView.
 Various CTRL+D,E and consistency fixes.
